refactor(parser): split primary/regex.rs into cohesive submodules (§7-8)
Split the 2385-line parser/primary/regex.rs (over the 500-line limit) into a
facade plus six themed submodules under src/parser/primary/regex/. regex_lit
is a single ~800-line indivisible parser (handles all Raku regex literal
forms, adverbs, m//, rx//, tr///, s///, version, topic-method), so lit.rs is
intentionally over 500 lines. scan_to_delim_inner (~250 lines) is the other
indivisible large function.

- regex.rs (30)        - facade: mod decls + pub(in crate::parser) re-exports
- adverbs.rs (415)     - MatchAdverbs struct + all adverb parse/validate helpers
- call_args.rs (155)   - parse_call_arg_list + colon-method-arg helpers
- lit.rs (1248)        - regex_lit / version_lit / topic_method_call (3 indivisible parsers)
- scan.rs (302)        - scan_to_delim / scan_to_delim_p5 / scan_to_delim_inner
- subst.rs (259)       - s/// replacement expression + topic-subst builder helpers
- trans.rs (71)        - tr/// escape processing + adverb parsing

Pure relocation. Submodule items are pub(super) or pub(in crate::parser);
re-exported at original visibility from the facade. version_lit and
topic_method_call bumped to pub(in crate::parser::primary) so the facade can
re-export them as pub(super). super::/super::super:: paths rewritten to
crate::parser:: absolute paths. External callers (primary/mod.rs, ident.rs,
comparison.rs, token_body.rs, reduction.rs) are untouched. The fn multiset is
byte-identical. No behavior change.
